+++ test/tap/control/command_scheduler_tests.cpp
@@ -601,13 +601,14 @@ TEST(CommandScheduler, runHardwareTest_does_not_segfault_with_nullptr_command)
     CommandScheduler scheduler(&drivers, true);
 
     SubsystemMock s1(&drivers);
-    EXPECT_CALL(s1, getTestCommand).WillOnce(Return(nullptr));
+    EXPECT_CALL(s1, getTestCommand).WillRepeatedly(Return(nullptr));
     EXPECT_CALL(s1, refresh);
     EXPECT_CALL(s1, getDefaultCommand).WillOnce(Return(nullptr));
 
     scheduler.registerSubsystem(&s1);
     scheduler.runHardwareTest(&s1);
     scheduler.run();
+    EXPECT_FALSE(scheduler.hasPassedTest(&s1));
 }
 
 TEST(CommandScheduler, runHardwareTest_runs_test_command)
@@ -628,14 +629,15 @@ TEST(CommandScheduler, runHardwareTest_runs_test_command)
 
     // Called once in the run loop.
     EXPECT_CALL(c1, execute).Times(1);
-    EXPECT_CALL(c1, isFinished).WillOnce(Return(true));
+    EXPECT_CALL(c1, isFinished).Times(2).WillRepeatedly(Return(true));
     EXPECT_CALL(c1, end(false)).Times(1);
 
     scheduler.registerSubsystem(&s1);
     scheduler.runHardwareTest(&s1);
-    EXPECT_TRUE(scheduler.runningTest(&s1));
+    EXPECT_TRUE(scheduler.isRunningTest(&s1));
     EXPECT_EQ(scheduler.countRunningHardwareTests(), 1);
     scheduler.run();
+    EXPECT_TRUE(scheduler.hasPassedTest(&s1));
 }
 
 TEST(CommandScheduler, runHardwareTest_then_stop_initializes_then_ends_test_command)
@@ -662,6 +664,7 @@ TEST(CommandScheduler, runHardwareTest_then_stop_initializes_then_ends_test_comm
     scheduler.runHardwareTest(&s1);
     scheduler.stopHardwareTest(&s1);
     scheduler.run();
+    EXPECT_FALSE(scheduler.hasPassedTest(&s1));
 }
 
 TEST(CommandScheduler, runAllHardwareTests_runs_all_valid_test_commands)
@@ -682,7 +685,7 @@ TEST(CommandScheduler, runAllHardwareTests_runs_all_valid_test_commands)
 
     // Called once in the run loop.
     EXPECT_CALL(c1, execute).Times(1);
-    EXPECT_CALL(c1, isFinished).WillOnce(Return(true));
+    EXPECT_CALL(c1, isFinished).Times(2).WillRepeatedly(Return(true));
     EXPECT_CALL(c1, end(false)).Times(1);
 
     SubsystemMock s2(&drivers);
@@ -694,6 +697,8 @@ TEST(CommandScheduler, runAllHardwareTests_runs_all_valid_test_commands)
     scheduler.registerSubsystem(&s2);
     scheduler.runAllHardwareTests();
     scheduler.run();
+    EXPECT_TRUE(scheduler.hasPassedTest(&s1));
+    EXPECT_FALSE(scheduler.hasPassedTest(&s2));
 }
 
 TEST(CommandScheduler, stopAllHardwareTests_stops_all_valid_test_commands)
@@ -726,6 +731,7 @@ TEST(CommandScheduler, stopAllHardwareTests_stops_all_valid_test_commands)
     scheduler.runAllHardwareTests();
     scheduler.stopAllHardwareTests();
     scheduler.run();
+    EXPECT_FALSE(scheduler.hasPassedTest(&s1));
 }
 
 TEST(CommandScheduler, runHardwareTest_cancels_scheduled_command_and_then_runs_test_command)
@@ -739,17 +745,16 @@ TEST(CommandScheduler, runHardwareTest_cancels_scheduled_command_and_then_runs_t
     set<Subsystem *> requirements = {&s1};
     NiceMock<CommandMock> c1;
     EXPECT_CALL(c1, getRequirementsBitwise)
-        .Times(4)  // Called when adding and removing the command
         .WillRepeatedly(Return(calcRequirementsBitwise(requirements)));
     EXPECT_CALL(c1, initialize).Times(2);
-    EXPECT_CALL(s1, getDefaultCommand).Times(2).WillRepeatedly(Return(&c1));
+    EXPECT_CALL(s1, getDefaultCommand).WillRepeatedly(Return(&c1));
 
     NiceMock<CommandMock> c2;
     EXPECT_CALL(c2, getRequirementsBitwise)
         .Times(2)  // Called when adding and removing the command
         .WillRepeatedly(Return(calcRequirementsBitwise(requirements)));
     EXPECT_CALL(c2, initialize);
-    EXPECT_CALL(s1, getTestCommand).WillOnce(Return(&c2));
+    EXPECT_CALL(s1, getTestCommand).WillRepeatedly(Return(&c2));
 
     // Called once in the second run().
     EXPECT_CALL(c1, execute).Times(1);
@@ -758,7 +763,7 @@ TEST(CommandScheduler, runHardwareTest_cancels_scheduled_command_and_then_runs_t
 
     // Called once in the third run().
     EXPECT_CALL(c2, execute).Times(1);
-    EXPECT_CALL(c2, isFinished).WillOnce(Return(true));
+    EXPECT_CALL(c2, isFinished).Times(3).WillRepeatedly(Return(true));
     EXPECT_CALL(c2, end(false)).Times(1);
 
     scheduler.registerSubsystem(&s1);
@@ -766,6 +771,8 @@ TEST(CommandScheduler, runHardwareTest_cancels_scheduled_command_and_then_runs_t
     scheduler.run();                 // Runs the default command
     scheduler.runHardwareTest(&s1);  // Add the test command and cancel the default command
     scheduler.run();  // Run the test command, finish it, and schedule the default command
+    EXPECT_TRUE(scheduler.hasPassedTest(&s1));
+    EXPECT_TRUE(scheduler.isCommandScheduled(&c1));
 }
 
 TEST(CommandScheduler, run_with_single_registered_subsystem_calls_refresh)